requirements.txt
uvicorn
python-dotenv
google-generativeai
pyPDF2
python-multipart
streamlit